Refactor: 피드백 반영 - 패키지 구조, ExceptionHandler
Application 파일을 제외한 파일들을 다른 패키지로 이동.
ExceptionHandler에서 @ResponseStatus 어노테이션을 활용하여
간단하게 response code를 반환하도록 변경.
