+++ src/main/java/subway/config/AuthenticationPrincipalConfig.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
