+++ src/main/java/subway/config/DataLoader.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.config;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Profile;
