+++ src/main/java/subway/config/SubwayExceptionHandler.java
@@ -1,18 +1,17 @@
-package subway;
+package subway.config;
 
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import subway.auth.exception.AuthException;
 
 @RestControllerAdvice
-public class SubwayAdvice {
+public class SubwayExceptionHandler {
 
     @ExceptionHandler(AuthException.class)
-    public ResponseEntity<Void> handleAuthException(Exception e) {
+    @ResponseStatus(value = HttpStatus.UNAUTHORIZED)
+    public void handleAuthException(Exception e) {
         e.printStackTrace();
-
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 }
